Include nocmodl build, nocmodl tests
Testing nocmodl output requires pre-filtering
of outputs and validation files, as the
compiler introduces a hard-coded path.

Build *does not install* nocmodl. This is
to avoid conflicts with HPCNeuron. This
potential ambiguity may be avoided in
the future by renaming nocmodl and nocmodl_core
as produced by sim/mod2c to mod2c and mod2c_core
respectively.

Change-Id: Ic17fdb39f292365561b22fcbfe8350d6565bf02c