Bugfix: heap corruption on mechanism reg [CBLUR-57]
Mechanisms that are compiled in and registered with mod_func.c
but are not listed in bb_mech.dat can scribble on the heap
when their mechanism type lookup fails.

This patch adds a check for this, and modules will return
without registering if they are not present in the run-time
mechanism table.

Amend: update validation .c files to suit.

Amend: incorporated Michael Hines' recommendations.

Change-Id: I6e2a11121bd5573e53b489371644a4083558f526